Renames package's folder name to `stringutils`
