+++ core/utils/stringutils/string_utils.go
