+++ core/utils/stringutils/string_utils_test.go
